Move BuildDefaultStorageFactory to kubeapiserver
