+++ pkg/kubeapiserver/BUILD
@@ -5,12 +5,24 @@ licenses(["notice"])
 load(
     "@io_bazel_rules_go//go:def.bzl",
     "go_library",
+    "go_test",
 )
 
 go_library(
     name = "go_default_library",
-    srcs = ["doc.go"],
+    srcs = [
+        "default_storage_factory_builder.go",
+        "doc.go",
+    ],
     tags = ["automanaged"],
+    deps = [
+        "//pkg/api:go_default_library",
+        "//pkg/genericapiserver:go_default_library",
+        "//pkg/storage/storagebackend:go_default_library",
+        "//pkg/util/config:go_default_library",
+        "//vendor:k8s.io/apimachinery/pkg/runtime",
+        "//vendor:k8s.io/apimachinery/pkg/runtime/schema",
+    ],
 )
 
 filegroup(
@@ -30,4 +42,20 @@ filegroup(
         "//pkg/kubeapiserver/options:all-srcs",
     ],
     tags = ["automanaged"],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = ["default_storage_factory_builder_test.go"],
+    library = ":go_default_library",
+    tags = ["automanaged"],
+    deps = [
+        "//pkg/api:go_default_library",
+        "//pkg/api/install:go_default_library",
+        "//pkg/api/v1:go_default_library",
+        "//pkg/apis/extensions/install:go_default_library",
+        "//pkg/apis/extensions/v1beta1:go_default_library",
+        "//pkg/genericapiserver:go_default_library",
+        "//vendor:k8s.io/apimachinery/pkg/runtime/schema",
+    ],
 )
